requirements.txt
oci
paramiko
python-dotenv
pystray
pillow
requests